refactor: create form hook
